Refresh joined rooms when enabling crypto
When enabling crypto on an Intent, refresh its list of joined rooms,
as it's used to determine which rooms to watch for crypto updates.

This fixes an issue where a restart would cause Intents to lose track of
which rooms they were joined to, until the list is refreshed either
manually or in response to a membership event.

Signed-off-by: Andrew Ferrazzutti <andrewf@element.io>